requirements-dev.txt
pytest-django
blue >= "0.7.0"
isort >= ^5.10.1"
mkdocs >= "1.2.3"
prospector >= "1.6.0"
pip-audit >= "1.1.2

-r requirements.txt